+++ scripts/luarocks_clean.sh
@@ -1,11 +1,14 @@
 #!/usr/bin/env bash
 
 # clean up Lua Rocks from global library
-luarocks remove --global argmap-3.1.2-3.rockspec
+luarocks remove --global argmap-3.2.0-4.rockspec
+luarocks remove argmap
 luarocks remove lualogging
 luarocks remove lyaml
 luarocks remove penlight
 luarocks remove rxi-json-lua
 luarocks remove luasocket
 luarocks remove luafilesystem
+
+echo "Remaining luarocks:"
 luarocks list