+++ scripts/tests.sh
@@ -2,8 +2,18 @@
 
 echo 'Attempting to delete old test outputs.'
 
-rm "Output/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
-rm "Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
+TEST_FILE_YML=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml
+TEST_FILE_MUP=$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup
+TEST_FILE_MD=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.md
+
+rm "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
+rm "$WORKSPACE/Output/example.html"
+rm "$TEST_FILE_MUP"
+rm "$MJS_WP_MAP"
+
+# todo Delete old gdrive file
+# 1uU7_yfAwMPV3a0lxpiXoVR-m0hbX2Pzs
+# Though may not be consistently same name anyway, would need to create with fix name
 
 COLOUR='true' # true
 TESTNUM=1
@@ -43,13 +53,6 @@ test() {
     TESTNUM=$((TESTNUM + 1))
 }
 
-# todo Delete old gdrive file
-# 1uU7_yfAwMPV3a0lxpiXoVR-m0hbX2Pzs
-# Though may not be consistently same name anyway, would need to create with fix name
-
-TEST_FILE_YML="Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
-TEST_FILE_MUP="Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
-
 # todo turn these into an array
 # array0=(one two three four five six)
 # for i in "${array0[@]}"; do
@@ -65,10 +68,12 @@ TEST_FILE_MUP="Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
 # do
 #     cat "$Output"
 # done
-test a2m $TEST_FILE_YML
-test m2a $TEST_FILE_MUP
-test a2mu $TEST_FILE_YML
-test a2mo $TEST_FILE_YML
+test a2m "$TEST_FILE_YML"
+test m2a "$TEST_FILE_MUP"
+test a2t "$TEST_FILE_YML"
+test a2mu "$TEST_FILE_YML"
+test a2mo "$TEST_FILE_YML"
+test md2h "$TEST_FILE_MD"
 
 echo "Testing finished, $FAILCOUNT tests failed."